Add Board Preference setting for puzzle screens
